Updated Unity Project assets and Unity version to 2020.3 LTS
