+++ UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLab-F_CA01.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 178d73691ae7249439aa3d8b33b246ff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
