+++ UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLab-F_CA01/F_CA01 Variant.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b46cef8414e6347d397a699c573125a0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
