+++ UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLab-M_CA01.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ad729f882621413f96bb0d548aff55e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
