+++ UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLab-M_CA01/M_CA01 Variant.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 72c34fad5b6d245beab2585368f8a7db
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
