--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 16dc193458937454595e06ce9375ef22
-folderAsset: yes
-timeCreated: 1529341629
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
