--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_cornea1528896426_737091.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 27799a872cda3419396e42c7af42ad4e
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
