--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_fur1528896426_737072.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 79c609dc5162d4df68d86828785d246e
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
