--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_generic1528896426_737078.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 3840026c72c0845f48323d1a84062f00
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
