--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_human_eyes1528896426_737085.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ddc8d6bad8e024bdb93308df6372274b
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
