--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_human_skin1528896426_737097.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: f90b130c57ce04adaa42cec4335af7ef
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
