--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_human_teeth1528896426_737064.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 5c0a0afc1eb584a26800d52d85d753a2
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
