--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabFemale/MBlab_pupil1528896426_73704.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 558c0bf34f053403a909105d433c3d40
-timeCreated: 1529341785
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
