--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 56817caf34212403183a766fd56af781
-folderAsset: yes
-timeCreated: 1529508280
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
