--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_cornea1528898645_097982.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 286e08dbce2ac4c8b89ea8c556d64c08
-timeCreated: 1529509047
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
