--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_fur1528898645_097965.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 413982ea2c39945448cbf97effc7bfd6
-timeCreated: 1529509047
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
