--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_generic1528898645_09797.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 9777c55f842ec47a7b695e65e3d37d8e
-timeCreated: 1529509047
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
