--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_human_eyes1528898645_097976.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 8905cad1901d540bbb9036448b4f9995
-timeCreated: 1529509048
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
