--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_human_skin1528898645_097986.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a228ec8d216ea47bda54e8f1160f0225
-timeCreated: 1529509048
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
