--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_human_teeth1528898645_097958.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ff26dd9d6e47f4453a95f83a606fe3d7
-timeCreated: 1529509048
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
