--- UnityProjects/YallahTestbed/Assets/YALLAH/Characters/MBLabMale/MBlab_pupil1528898645_097935.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 3ba946a6959bf493da15f13c57b60ba7
-timeCreated: 1529509048
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
