--- UnityProjects/YallahTestbed/Assets/YALLAH/Scenes/ScriptsForTesting/AllFeaturesTester.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c8aa10a1d8fb74ad5926f4ba4927bb82
-timeCreated: 1506504713
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
