--- UnityProjects/YallahTestbed/Assets/YALLAH/Scenes/TestEyeHeadGaze.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4061939372a6a40ef82103abee7e8c45
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
