--- UnityProjects/YallahTestbed/Assets/YALLAH/Scenes/TestFemaleBase.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 4ee9a0d482ecc4b40bb460513c6cf0b5
-timeCreated: 1529341862
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
