--- UnityProjects/YallahTestbed/Assets/YALLAH/Scenes/TestMaleBase.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 1e6886a158c7249c28eadbb0af9e9184
-timeCreated: 1529331621
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
