+++ UnityProjects/YallahTestbed/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,2 @@
-m_EditorVersion: 2017.4.14f1
+m_EditorVersion: 2020.3.14f1
+m_EditorVersionWithRevision: 2020.3.14f1 (d0d1bb862f9d)
